Refactoring wallets endpoints
