+++ src/cryptoadvance/specter/server_endpoints/controller.py
@@ -41,7 +41,7 @@ from .price import price_endpoint
 from .settings import settings_endpoint
 from .setup import setup_endpoint
 from .wallets import wallets_endpoint
-from .wallets_api import wallets_endpoint_api
+from .wallets.wallets_api import wallets_endpoint_api
 from .welcome import welcome_endpoint
 
 spc_prefix = app.config["SPECTER_URL_PREFIX"]
