+++ src/cryptoadvance/specter/server_endpoints/wallets/__init__.py
@@ -1,0 +1,1 @@
+from .wallets import *
