+++ src/cryptoadvance/specter/server_endpoints/wallets/wallets_api.py
@@ -20,19 +20,19 @@ from flask_babel import lazy_gettext as _
 from flask_login import current_user, login_required
 from werkzeug.wrappers import Response
 
-from ..commands.psbt_creator import PsbtCreator
-from ..helpers import bcur2base64
-from ..rpc import RpcError
-from ..server_endpoints import flash
-from ..server_endpoints.filters import assetlabel
-from ..specter_error import SpecterError, handle_exception
-from ..util.base43 import b43_decode
-from ..util.descriptor import Descriptor
-from ..util.fee_estimation import FeeEstimationResultEncoder, get_fees
-from ..util.mnemonic import generate_mnemonic
-from ..util.price_providers import get_price_at
-from ..util.tx import decoderawtransaction
-from ..wallet import Wallet
+from ...commands.psbt_creator import PsbtCreator
+from ...helpers import bcur2base64
+from ...rpc import RpcError
+from ...server_endpoints import flash
+from ...server_endpoints.filters import assetlabel
+from ...specter_error import SpecterError, handle_exception
+from ...util.base43 import b43_decode
+from ...util.descriptor import Descriptor
+from ...util.fee_estimation import FeeEstimationResultEncoder, get_fees
+from ...util.mnemonic import generate_mnemonic
+from ...util.price_providers import get_price_at
+from ...util.tx import decoderawtransaction
+from ...wallet import Wallet
 
 logger = logging.getLogger(__name__)
